move tests to correct place
